Remove Field-Level IsReflected Flag
Following the removal of the Reflected annotation it no longer makes sense to identify whether a field is reflected when we can check whether a type is reflected. The type of the field could indeed be reflected, or a type inside a container could be, a map key could be and the map value not (or vise versa), or different tuple elements could be reflected or not... Whether something is reflected is ambiguous at the field level and needs to be checked at the type level.